Add 204-byte BTS packet support
- Add TS_PACKET_SIZE_204 and TS_RS_PARITY_SIZE constants

- Auto-detect packet format in TSDemuxer using chunk length modulo

- Strip 16-byte RS parity suffix when demuxing 204-byte packets

- Replace NSAssert crashes with graceful error handling (log + return)

- Fix pcrIntervalSeconds type from NSUInteger to double (was truncating 0.04)

- Migrate tests from Swift to Objective-C

+++ Sources/TSMuxDemux/TSMuxer.m
@@ -235,7 +235,7 @@
     if (self.pcrPid != accessUnit.pid) {
         return 0;
     }
-    static const NSUInteger pcrIntervalSeconds = 0.04;
+    static const double pcrIntervalSeconds = 0.04;
     
     const BOOL hasSentPcr = CMTIME_IS_VALID(self.firstSentPcr);
     const Float64 secondsElapsedSinceLastPcr = CMTimeGetSeconds(accessUnit.pts) - CMTimeGetSeconds(self.lastSentPcr);
